Move PSI Setup Success Notification to notifications directory.
