+++ assets/js/modules/pagespeed-insights/components/notifications/SetupSuccessNotification.js
